Bug fixes - refactored intersectionObserver to mixin
